Rename "item" folder to "demo".
