+++ demo-angular/app/demo/standard.component.html
