+++ demo-angular/app/demo/standard.component.ts
